Box2D System + FixedTimeStep
Tombstone objects are now taken into account when adding Objects.
Resolves Issue #2.

Created a Box2D system that contains a Hash which allows for quick
lookup of a Box2D body given an arbitrary GameObject.
Resolves #1

Mostly addresses #3, however, we still aren't doing anything with the
time left in the accumulator at the end of the Frame. This is probably
OK for now.